start on new display
I had to start with the new way Its going to display the boards before I can do the whole attack thing. I started with it building the player's board.
I've yet to test it but I need to get back to work so I'll have to leave it here for now.

Don't forget to add a part of the function that checks to see if the ship is hit!! There should be another for that loops through the ships. or in the same one to make a mark for it they were hit at one point.