basic assign questions loop
I realized as I was doing this that I could do it in a loop to make it so I don't have to write out this code each and every time I want the ships to be assigned. When I come back to this I need to make it so that when the user puts the ship in a place that isn't good that the loop will restart and make them try again.