test function
testing brought a few bugs into the light. I fixed one where it didn't like that the user input for number was a string. But now It's saying that the ship can't be placed when it can. So I need to look into the actual assign function for that one.